backend/requirements.txt
# Core API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML Core
torch>=2.0.0
torchvision
torchaudio
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
regex>=2023.0.0

# Graph Neural Network (PyTorch Geometric)
torch-geometric
networkx>=3.0

# Transformers (GraphCodeBERT)
transformers>=4.35.0
sentencepiece

# Gemini AI (new SDK)
google-genai>=1.0.0

# Static Analysis
# slither-analyzer  # Uncomment after: pip install slither-analyzer

# Utilities
tqdm>=4.65.0
requests>=2.31.0
